Implement a sample db that can be accessed everywhere in the App.
