+++ WpfApp1/Dashboard.xaml.cs
@@ -23,15 +23,10 @@ namespace WpfApp1
     {
         public Dashboard()
         {
-            List<Appointment> appointments = new List<Appointment>();
-            appointments.Add(new Appointment("Rupert", "Amodia", 1, "9:00AM", "9:30AM", "Dr. Chirag"));
-            //appointments.Add(new Appointment("Araiz", "Asad", 1, "10:00AM", "10:30AM", "Dr. Raphael"));
-            //appointments.Add(new Appointment("Elizabeth Chu", "Asad", 1, "11:00AM", "10:30AM", "Dr. Amr"));
-            //appointments.Add(new Appointment("David", "Smith", 1, "11:00AM", "11:30AM", "Dr. Raphael"));
-            //appointments.Add(new Appointment("John", "Cena", 1, "12:30PM", "1:00PM", "Dr. Raphael"));
-            //appointments.Add(new Appointment("Matthew", "Murdock", 1, "1:00PM", "1:30PM", "Dr. Amr"));
-            //appointments.Add(new Appointment("Jennifer", "Walters", 1, "1:00PM", "1:30PM", "Dr. Chirag"));
-            //appointments.Add(new Appointment("Muhammad", "Mohammed", 1, "2:00PM", "2:30PM", "Dr. Chirag"));
+            List<Appointment> appointments = DB.Appointments;
+
+            // say we want to add an appointment to our 'DB'
+            appointments.Add(new Appointment("New", "Patient", 1, "3:00PM", "3:30PM", "Dr. Chirag"));
 
             InitializeComponent();
             quickView.ItemsSource = appointments;
