+++ WpfApp1/MainWindow.xaml.cs
@@ -25,83 +25,11 @@ namespace WpfApp1
         public MainWindow()
         {
             InitializeComponent();
-
-            MainNavWindow.Source = new Uri("Calendar.xaml", UriKind.Relative);
-
+            MainNavWindow.Source = new Uri("Dashboard.xaml", UriKind.Relative);
 
 
-        }
-
-        public class Appointment
-        {
-            private string _firstName;
-            private string _lastName;
-            private int _day;
-            private string _start;
-            private string _end;
-            private string _doctor;
 
 
-            public string FirstName
-            {
-                get { return _firstName; }
-                set { _firstName = value; }
-            }
-
-            public string LastName
-            {
-                get { return _lastName; }
-                set { _lastName = value; }
-            }
-
-            public int Day
-            {
-                get { return _day; }
-                set { _day = value; }
-            }
-
-            public string Start
-            {
-                get { return _start; }
-                set { _lastName = value; }
-            }
-
-            public string End
-            {
-                get { return _end; }
-                set { _lastName = value; }
-            }
-
-            public string StartEnd
-            {
-                get { return _start + " - " + _end; }
-            }
-
-            public string FullName
-            {
-                get { return _firstName + " " + _lastName;  }
-            }
-
-            public string Doctor
-            {
-                get { return _doctor; }
-            }
-
-            public Appointment(
-                string firstname,
-                string lastname,
-                int day,
-                string start,
-                string end,
-                string doctor)
-            {
-                _firstName = firstname;
-                _lastName = lastname;
-                _day = day;
-                _start = start;
-                _end = end;
-                _doctor = doctor;
-            }
         }
     }
 }
